Implement GetOrders query with pagination and mock Domain Event handler

+++ src/Services/Ordering/Ordering.Application/Dtos/OrderItemDto.cs
@@ -5,5 +5,5 @@ public class OrderItemDto
     public Guid OrderId { get; set; }
     public Guid ProductId { get; set; }
     public int Quantity { get; set; }
-    public int Price { get; set; }
+    public decimal Price { get; set; }
 }
